+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -539,7 +539,8 @@ public final class CollectionsPlume {
     "signedness", // problem with clone()
     "nullness" // generics problem
   })
-  public static <T extends @Nullable Object, C extends @Nullable Collection<T>> @PolyNull C cloneElements(@PolyNull C orig) {
+  public static <T extends @Nullable Object, C extends @Nullable Collection<T>>
+      @PolyNull C cloneElements(@PolyNull C orig) {
     if (orig == null) {
       return null;
     }
@@ -563,7 +564,8 @@ public final class CollectionsPlume {
    * @return a copy of {@code orig}, as described above
    */
   @SuppressWarnings({"signedness", "nullness:argument"}) // problem with clone()
-  public static <T extends @Nullable DeepCopyable<T>, C extends @Nullable Collection<T>> @PolyNull C deepCopy(@PolyNull C orig) {
+  public static <T extends @Nullable DeepCopyable<T>, C extends @Nullable Collection<T>>
+      @PolyNull C deepCopy(@PolyNull C orig) {
     if (orig == null) {
       return null;
     }
@@ -624,6 +626,20 @@ public final class CollectionsPlume {
       }
     }
     return result;
+  }
+
+  /**
+   * Returns the first index of the given value in the list, starting at the given index. Uses
+   * {@code Object.equals()} for comparison.
+   *
+   * @param list a list
+   * @param value the value to search for
+   * @param start the starting index
+   * @return the index of the value in the list, at or after the given index
+   */
+  public static int indexOf(List<?> list, Object value, int start) {
+    int idx = list.subList(start, list.size()).indexOf(value);
+    return idx == -1 ? -1 : idx + start;
   }
 
   ///////////////////////////////////////////////////////////////////////////
@@ -1559,7 +1575,8 @@ public final class CollectionsPlume {
    * @return a copy of {@code orig}, as described above
    */
   @SuppressWarnings({"nullness", "signedness"}) // generics problem with clone
-  public static <K, V extends @Nullable DeepCopyable<V>, M extends @Nullable Map<K, V>> @PolyNull M deepCopyValues(@PolyNull M orig) {
+  public static <K, V extends @Nullable DeepCopyable<V>, M extends @Nullable Map<K, V>>
+      @PolyNull M deepCopyValues(@PolyNull M orig) {
     if (orig == null) {
       return null;
     }